Developed user crud operations
Developed user CRUD operations. Created the AddUser component for adding new users, the ViewUser component for viewing user details, and the UpdateUser component for updating existing users. Integrated Redux for state management of user data and added necessary handlers and state management in UserPage. Applied Tailwind CSS for styling the components and ensured responsive design for all components. Added routing to display the user page when the user button in the sidebar is clicked.